FEBULIA.jl

Documentation for FEBULIA.jl

FEBULIA.FEBULIA — Module

This is the FEBULIA, it contains

FEBULIA.BoundCond1D

FEBULIA.riemann

FEBULIA.dotf

FEBULIA.compute_norm

FEBULIA.coefficient

FEBULIA.PolyExp

FEBULIA.shift_PolyExp

FEBULIA.evalPolyExp

FEBULIA.PolyExpBasisFun

FEBULIA.deriv

FEBULIA.polynomial_primitive

FEBULIA.polynomial_deriv

FEBULIA.integrate

FEBULIA.basis_PE_BC

FEBULIA.basis_PE

FEBULIA.basis

FEBULIA.BoundCond1D — Type

BoundCond1D 

This type contains all the necessary information related to the boundary conditions of a FE problem in 1D

 - BCl and BCu: strings describing the boundary conditions (valid type in {"Dirichlet", "Neumann", "Robin"})
 - xl and xu: are the boundaries (xl: low, xu: up)
 - ul and uu: values of the boundary conditions
 - g: a function that gives the values at the boundaries, it may vary with an external parameter which is why it is a function
 - a and b: describe the Robin conditions ($a u(x) + b \frac{\partial u}{\partial x}(x) = g(x)$)

 An object is constructed with the function
    
    BoundCond1D(BCl::String,BCu::String,xl::Cdouble,xu::Cdouble;lowBC::Cdouble=0.0,upBC::Cdouble=0.0,Ra::Cdouble=0.0,Rb::Cdouble=0.0,Rg::Function=(y::Cdouble->0.0))

FEBULIA.basis_PE_BC — Function

basis_PE_BC(X::Array{Cdouble,1},p1::PolyExp,p2::PolyExp,BC::BoundCond1D)

generates a `basis_PE` object from

  - X:         array with N discretization nodes 
  - p1 and p2: generator PolyObjects (p1 for the first part of the interval of the basis function and p2 for the second) These PolyExp objects are shifted along the discretization nodes
  - BC:        boundary conditions. According to the type of bnoundary conditions, the basis can have N, N-1 or N-2 basis function

FEBULIA.basis_PE — Type

basis_PE

a structure that represents basis functions with `PolyExp` objects 

  - N: number of basis functions
  - xl, xm and xu: arrays containing the lower, intermediate, and upper end of the support of the functions (each of these array have N elements)
  - p1 and p2: arrays with N PolyExp object representing the basis functions over the respective intervals [xl,xm] and [xm,xu]
  - v: array of N basis functions 

A basis_PE object should be created with the function

basis_PE(xl::Array{Cdouble,1},xm::Array{Cdouble,1},xu::Array{Cdouble,1},p1::Array{PolyExp,1},p2::Array{PolyExp,1},v::Array{Function,1})

FEBULIA.basis — Type

basis

WARNING: soon deprecated

a structure that contains basis functions 

  - N: number of basis functions
  - X: 2D array containing the lower and upper end of the support of the functions (dim=1: basis function index, dim=2: lower and upper interval ends)
  - v: array of N basis functions 

A basis_PE object should be created with the function

basis_PE(xl::Array{Cdouble,1},xm::Array{Cdouble,1},xu::Array{Cdouble,1},p1::Array{PolyExp,1},p2::Array{PolyExp,1},v::Array{Function,1})