Inner products

FEBULIA.riemann — Function

riemann(f::Function, a::Real, b::Real, n::Int; method="right") 

numerical quadrature to compute the integral value of the function f over the interval [a,b] using n nodes

riemann(fs::Array{Cdouble,1},xs::Array{Cdouble,1}; method="trapezoid")

computes the numerical integration of the function f whose values are passed in fs for the nodes xs (fs = f.(xs))

optional argument

    - method ∈ {"right","left","trapezoid","simpsons"}

    "right":     f(r) * (r-l)
    "left":      f(l) * (r-l)
    "trapezoid": (1/2) * (f(l) + f(r)) * (r-l)
    "simpsons":  (1.0/6.0) * (f(l) + 4.0*(f((l+r)/2.0)) + f(r)) * (r-l)

FEBULIA.dotf — Function

dotf(f::Function,g::Function,x_min::Cdouble,x_max::Cdouble;N::Int64=10000)
dotf(f::Array{Cdouble,1},g::Array{Cdouble,1},x::Array{Cdouble,1})

Dot product for the functions f and g over the interval [x_min,x_max]

    (f,g) = ∫ f(x)g(x) dx

the functions f and g can be passed as argument and will then be evaluated in N nodes in the interval [x_min,x_max]
or the values can be passed in arrays f = f(x) and g = g(x)

FEBULIA.compute_norm — Function

compute_norm(f::Function,x_min::Cdouble,x_max::Cdouble)
compute_norm(fs::Array{Cdouble,1},xs::Array{Cdouble,1})

Norm of the function f over the interval [x_min,x_max]

    ||f|| = √(∫(f(x))^2 dx)

the function f can be passed as argument and will then be evaluated in N nodes in the interval [x_min,x_max]
or the values can be passed in an array f = f(x)

FEBULIA.coefficient — Function

coefficient(B::basis,f::Function)

computes the coefficients of the projection of the function f onto the basis function of B 
    c_i = (2/(Δx)) * ∫ f(x)*B.v[i](x) dx