PolyExp

FEBULIA.PolyExp — Type

PolyExp

Object that represents exponential-polynomials of the form ``P(X) = e^{\alpha X}\underset{n=0}{\overset{N}{\sum}} c_{N-n+1} X^n``

An object can be created with PolyExp(nn::Int64,cc::Array{Cdouble,1},``\alpha``::Cdouble)

``c = [1.0; 2.0; 3.0; 4.0]``
``\alpha = 0.7``
N = 3

p = PolyExp(N,c,``\alpha``)

PolyExp object can be multiplied * and compare ==

FEBULIA.shift_PolyExp — Function

shift_PolyExp(P::PolyExp,x0::Cdouble,x1::Cdouble;rev::Bool=false)

a function that computes the coefficient of a PolyExp when evaluated either in $rac{X-x_0}{x_1-x_0}$ (rev=false) or in $rac{x_1-X}{x_1-x_0}$ (rev=true)

The returned PolyExp P_shift(X) is the same degree as the input object

FEBULIA.evalPolyExp — Function

evalPolyExp(x::Cdouble,p::PolyExp)
evalPolyExp(x::Array{Cdouble,1},p::PolyExp)

evaluates the PolyExp for the value x

FEBULIA.PolyExpBasisFun — Function

PolyExpBasisFun(x::Cdouble,xmin::Cdouble,xmed::Cdouble,xmax::Cdouble,PE1::PolyExp,PE2::PolyExp)

evaluates in x the function defined as
```
\begin{cases}
    PE1(X) & \text{ if } X\in[x_{\min},x_{\text{med}})\\
    PE2(X) & \text{ if } X\in[x_{\text{med}},x_{\max})\\
    0 & \text{otherwise}
\end{cases}
```

FEBULIA.deriv — Function

deriv(PE::PolyExp)

creates a PolyExp object that is the derivative of PE, $rac{	ext{d}PE}{	ext{d}X}$

deriv(BPE::basis_PE)

returns a new `basis_PE` object with the derivatives of each function in BPE

FEBULIA.polynomial_primitive — Function

polynomial_primitive(PE::PolyExp)

creates a PolyExp object with the primitive of the PolyExp PE, i.e. the exponential coefficient is 0, the degree of the polynomial is PE.n+1, and the constant term is 0

FEBULIA.polynomial_deriv — Function

polynomial_deriv(PE::PolyExp)

creates a PolyExp object with the derivative of the PolyExp PE, i.e. the exponential coefficient is 0 and the degree of the polynomial is PE.n-1

FEBULIA.integrate — Function

integrate(PE::PolyExp,xmin::Cdouble,xmax::Cdouble)

compute recursively the integrale

I_n^{\alpha}(x_{\min},x_{\max}) = $\int_{x_{\min}}^{x_{\max}} e^{\alpha x} \underset{n=0}{\overset{N}{\sum}} c_{N-n+1} X^n \text{d}x$

if \alpha \neq 0, then 
    I_n^{\alpha}(x_{\min},x_{\max}) = \frac{1}{\alpha} \left(R_{n}^{\alpha}(x_{\min},x_{\max}) - I_{n-1}^{\alpha}(x_{\min},x_{\max})\right)
    I_0^{\alpha}(x_{\min},x_{\max}) = \frac{R_{0}^{\alpha}(x_{\min},x_{\max})}{\alpha}

    with R_{n}^{\alpha}(x_{\min},x_{\max}) = [\frac{\text{d}^{(N-n)} Poly(PE)}{\text{d} x}(x) e^{\alpha x}]_{x_{\min}}^{x_{\max}} 
    where Poly(PE) is the polynomial part of the PolyExp object, and by convention, the zero-th order derivative is the function

if \alpha=0, then

    I_n^{\alpha}(x_{\min},x_{\max}) = R_{n}^{\alpha}(x_{\min},x_{\max}) (as defined above)